seed departements

+++ Api/database/migrations/2024_09_12_095819_create_departments_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->string('id')->primary();
             $table->string("department_name");
             $table->timestamps();
-            $table->unsignedBigInteger('region_id');
+            $table->unsignedInteger('region_id');
             $table->foreign('region_id')->references('id')->on('regions');
         });
     }
